Refactor OpenAITeamAdapter to utilize ModelSettings for max_tokens and clean up OpenAISupervisorRepository by removing unused methods
