+++ inline_agents/backends/openai/backend.py
@@ -19,9 +19,7 @@ from inline_agents.backends.openai.sessions import (
     RedisSession,
     make_session_factory,
 )
-from nexus.inline_agents.backends.openai.models import (
-    OpenAISupervisor as Supervisor,
-)
+
 from nexus.inline_agents.models import InlineAgentsConfiguration
 from nexus.intelligences.models import ContentBase
 from nexus.projects.models import Project
@@ -30,49 +28,7 @@ from nexus.projects.websockets.consumers import (
 )
 from langfuse import get_client 
 from router.traces_observers.save_traces import save_inline_message_to_database
-
-
-class OpenAISupervisorRepository:
-    @classmethod
-    def get_supervisor(
-        cls,
-        project: Project,
-        foundation_model: str = None,
-    ) -> Agent:
-
-        supervisor = Supervisor.objects.order_by('id').last()
-
-        if not supervisor:
-            raise Supervisor.DoesNotExist()
-
-        supervisor_dict = {
-            "instruction": cls._get_supervisor_instructions(project=project, supervisor=supervisor),
-            "tools": cls._get_supervisor_tools(project=project, supervisor=supervisor),
-            "foundation_model": supervisor.foundation_model,
-            "knowledge_bases": supervisor.knowledge_bases,
-            "prompt_override_configuration": supervisor.prompt_override_configuration,
-            "default_instructions_for_collaborators": supervisor.default_instructions_for_collaborators,
-            "max_tokens": supervisor.max_tokens,
-        }
-
-        return supervisor_dict
-
-    @classmethod
-    def _get_supervisor_instructions(cls, project, supervisor) -> str:
-        if project.use_components and project.human_support:
-            return supervisor.components_human_support_prompt
-        elif project.use_components:
-            return supervisor.components_prompt
-        elif project.human_support:
-            return supervisor.human_support_prompt
-        else:
-            return supervisor.instruction
-
-    @classmethod
-    def _get_supervisor_tools(cls, project, supervisor) -> list[dict]:
-        if project.human_support:
-            return supervisor.human_support_action_groups
-        return supervisor.action_groups
+from nexus.inline_agents.backends.openai.repository import OpenAISupervisorRepository
 
 
 class OpenAIBackend(InlineAgentsBackend):
